Refactor common helpers
